+++ api/src/main/java/org/allaymc/api/utils/tuple/Pair.java
@@ -1,6 +1,8 @@
 package org.allaymc.api.utils.tuple;
 
 /**
+ * Pair is an immutable tuple of two values.
+ *
  * @author daoge_cmd
  */
 public record Pair<L, R>(L left, R right) {