+++ server/src/main/java/org/allaymc/server/container/AllayFakeContainerFactory.java
@@ -1,0 +1,21 @@
+package org.allaymc.server.container;
+
+import org.allaymc.api.container.FakeContainerFactory;
+import org.allaymc.api.container.interfaces.FakeContainer;
+import org.allaymc.server.container.impl.FakeChestContainerImpl;
+import org.allaymc.server.container.impl.FakeDoubleChestContainerImpl;
+
+/**
+ * @author daoge_cmd
+ */
+public class AllayFakeContainerFactory implements FakeContainerFactory {
+    @Override
+    public FakeContainer createFakeChestContainer() {
+        return new FakeChestContainerImpl();
+    }
+
+    @Override
+    public FakeContainer createFakeDoubleChestContainer() {
+        return new FakeDoubleChestContainerImpl();
+    }
+}